Bug Fixes
Found a bug where if you added a node to any tree then tried to remove it while the tree was empty it would crash remove but not init the root node back to a null value which would then cause it to throw null pointer exceptions.

SHOULD have fixed it. Not 100% sure but I did some tests and I think its fine