More bug fixes
Added a message for when a new manufacturer is created.
Changed one of the Input methods so when you enter an integer it trims the whitespace off the end of the string before attempting to parse to an integer. This was causing it to catch an exception every time it ran. Now its trimming that white space off and should be able to handle anything we throw at it